Add files via upload
Made a book app. Added a downloaded css file for styling in it. Made components - BookCreate, BookList, BookShow, BookEdit. BookCreate and BookList are the child components of App component. and BookShow and BookEdit are the child components of BookList component. Implemented CRUD operations in the application. Used "axios" for sending and receiving http requests and responses, and used "json server" to make the application persistent.